backend/django_requirements.txt
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
build==1.1.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.8
configparser==6.0.1
contourpy==1.2.0
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
dataclasses-json==0.6.4
datasets==2.18.0
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
diskcache==5.6.3
distlib==0.3.7
distro==1.9.0
Django==4.2.3
django-cors-headers==4.1.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.2.0
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.28.1
google-generativeai==0.4.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.1.2
ipykernel==6.29.3
ipython==8.22.2
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.10
langchain-community==0.0.25
langchain-core==0.1.33
langchain-google-genai==0.0.11
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.13
looseversion==1.3.0
lxml==5.1.0
Mako==1.3.5
Markdown==3.4.3
MarkupSafe==2.1.5
marshmallow==3.21.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
mmh3==4.1.0
monotonic==1.6
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.14.3
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
passlib==1.7.4
pathlib==1.0.1
Pillow==10.0.0
platformdirs==3.11.0
posthog==3.4.2
premailer==3.10.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
prov==2.0.0
psutil==5.9.8
psycopg2==2.9.6
pulsar-client==3.4.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic_core==2.16.3
pydot==2.0.0
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.1
pyparsing==3.1.2
pypdf==4.0.2
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pysbd==0.3.4
python-box==7.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2023.3
pywin32==306
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==25.1.2
ragas==0.1.5
rdflib==7.0.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tiktoken==0.6.0
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
traits==6.3.2
transformers==4.38.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.14.1
tzdata==2023.3
urllib3==2.2.1
uvicorn==0.27.1
virtualenv==20.24.5
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
whitenoise==6.9.0
wrapt==1.16.0
xxhash==3.4.1
yagmail==0.15.293
yarl==1.9.4
zipp==3.17.0
